+++ examples/index.php
@@ -107,6 +107,18 @@ require_once __DIR__ . '/../../../autoload.php';
                 <div class="card olive">
                     <div class="content">
                         <div class="header">
+                            Giropay
+                        </div>
+                        <div class="description">
+                        </div>
+                    </div>
+                    <div class="ui bottom attached green button" onclick="location.href='Giropay/';">
+                        Try
+                    </div>
+                </div>
+                <div class="card olive">
+                    <div class="content">
+                        <div class="header">
                             Alipay
                         </div>
                         <div class="description">
@@ -209,6 +221,18 @@ require_once __DIR__ . '/../../../autoload.php';
                 <div class="card olive">
                     <div class="content">
                         <div class="header">
+                            Flexipay (PIS)
+                        </div>
+                        <div class="description">
+                        </div>
+                    </div>
+                    <div class="ui bottom attached green button" onclick="location.href='FlexiPayDirect/';">
+                        Try
+                    </div>
+                </div>
+                <div class="card olive">
+                    <div class="content">
+                        <div class="header">
                             Hosted Payment Page
                         </div>
                         <div class="description">
@@ -252,7 +276,7 @@ require_once __DIR__ . '/../../../autoload.php';
                     <span class="sub header">Enable or disable webhooks ...</span>
                 </span>
             </h2>
-            <div class="ui two cards">
+            <div class="ui three cards">
                 <div class="card">
                     <div class="content">
                         <div class="header">
@@ -276,6 +300,19 @@ require_once __DIR__ . '/../../../autoload.php';
                         </div>
                     </div>
                     <div class="ui bottom attached blue button" onclick="location.href='Webhooks/removeAll.php';">
+                        Try
+                    </div>
+                </div>
+                <div class="card">
+                    <div class="content">
+                        <div class="header">
+                            Fetch all Webhooks
+                        </div>
+                        <div class="description">
+                            Fetch all Webhooks corresponding to this key pair.
+                        </div>
+                    </div>
+                    <div class="ui bottom attached blue button" onclick="location.href='Webhooks/fetchAll.php';">
                         Try
                     </div>
                 </div>