+++ src/Services/PaymentService.php
@@ -135,11 +135,13 @@ class PaymentService
      * @param Customer|string|null   $customer
      * @param string|null            $orderId
      * @param Metadata|string|null   $metadata
-     * @param Basket|null            $basket      The Basket object corresponding to the payment.
-     *                                            The Basket object will be created automatically if it does not exist
-     *                                            yet (i.e. has no id).
-     * @param bool|null              $card3ds     Enables 3ds channel for credit cards if available. This parameter is
-     *                                            optional and will be ignored if not applicable.
+     * @param Basket|null            $basket           The Basket object corresponding to the payment.
+     *                                                 The Basket object will be created automatically if it does not exist
+     *                                                 yet (i.e. has no id).
+     * @param bool|null              $card3ds          Enables 3ds channel for credit cards if available. This parameter is
+     *                                                 optional and will be ignored if not applicable.
+     * @param string|null            $invoiceId        The external id of the invoice.
+     * @param string|null            $paymentReference A reference text for the payment.
      *
      * @return Authorization Resulting Authorization object.
      *
@@ -155,7 +157,9 @@ class PaymentService
         $orderId = null,
         $metadata = null,
         $basket = null,
-        $card3ds = null
+        $card3ds = null,
+        $invoiceId = null,
+        $paymentReference = null
     ): AbstractTransactionType {
         $payment = $this->createPayment($paymentType);
         return $this->authorizeWithPayment(
@@ -167,7 +171,9 @@ class PaymentService
             $orderId,
             $metadata,
             $basket,
-            $card3ds
+            $card3ds,
+            $invoiceId,
+            $paymentReference
         );
     }
 
@@ -181,11 +187,13 @@ class PaymentService
      * @param Customer|string|null $customer
      * @param string|null          $orderId
      * @param Metadata|string|null $metadata
-     * @param Basket|null          $basket    The Basket object corresponding to the payment.
-     *                                        The Basket object will be created automatically if it does not exist
-     *                                        yet (i.e. has no id).
-     * @param bool|null            $card3ds   Enables 3ds channel for credit cards if available. This parameter is
-     *                                        optional and will be ignored if not applicable.
+     * @param Basket|null          $basket           The Basket object corresponding to the payment.
+     *                                               The Basket object will be created automatically if it does not exist
+     *                                               yet (i.e. has no id).
+     * @param bool|null            $card3ds          Enables 3ds channel for credit cards if available. This parameter is
+     *                                               optional and will be ignored if not applicable.
+     * @param string|null          $invoiceId        The external id of the invoice.
+     * @param string|null          $paymentReference A reference text for the payment.
      *
      * @return Authorization Resulting Authorization object.
      *
@@ -201,9 +209,14 @@ class PaymentService
         $orderId = null,
         $metadata = null,
         $basket = null,
-        $card3ds = null
+        $card3ds = null,
+        $invoiceId = null,
+        $paymentReference = null
     ): Authorization {
-        $authorization = (new Authorization($amount, $currency, $returnUrl))->setOrderId($orderId);
+        $authorization = (new Authorization($amount, $currency, $returnUrl))
+            ->setOrderId($orderId)
+            ->setInvoiceId($invoiceId)
+            ->setPaymentReference($paymentReference);
         if ($card3ds !== null) {
             $authorization->setCard3ds($card3ds);
         }
@@ -309,16 +322,18 @@ class PaymentService
     /**
      * Performs a Payout transaction and returns the resulting Payout resource.
      *
-     * @param float                  $amount      The amount to charge.
-     * @param string                 $currency    The currency of the amount.
-     * @param string|BasePaymentType $paymentType The PaymentType object or the id of the PaymentType to use.
-     * @param string                 $returnUrl   The URL used to return to the shop if the process requires leaving it.
-     * @param Customer|string|null   $customer    The Customer object or the id of the customer resource to reference.
-     * @param string|null            $orderId     A custom order id which can be set by the merchant.
-     * @param Metadata|null          $metadata    The Metadata object containing custom information for the payment.
-     * @param Basket|null            $basket      The Basket object corresponding to the payment.
-     *                                            The Basket object will be created automatically if it does not exist
-     *                                            yet (i.e. has no id).
+     * @param float                  $amount           The amount to charge.
+     * @param string                 $currency         The currency of the amount.
+     * @param string|BasePaymentType $paymentType      The PaymentType object or the id of the PaymentType to use.
+     * @param string                 $returnUrl        The URL used to return to the shop if the process requires leaving it.
+     * @param Customer|string|null   $customer         The Customer object or the id of the customer resource to reference.
+     * @param string|null            $orderId          A custom order id which can be set by the merchant.
+     * @param Metadata|null          $metadata         The Metadata object containing custom information for the payment.
+     * @param Basket|null            $basket           The Basket object corresponding to the payment.
+     *                                                 The Basket object will be created automatically if it does not exist
+     *                                                 yet (i.e. has no id).
+     * @param string|null            $invoiceId        The external id of the invoice.
+     * @param string|null            $paymentReference A reference text for the payment.
      *
      * @return Payout The resulting object of the Payout resource.
      *
@@ -333,7 +348,9 @@ class PaymentService
         $customer = null,
         $orderId = null,
         $metadata = null,
-        $basket = null
+        $basket = null,
+        $invoiceId = null,
+        $paymentReference = null
     ): AbstractTransactionType {
         $payment = $this->createPayment($paymentType);
         return $this->payoutWithPayment(
@@ -344,20 +361,25 @@ class PaymentService
             $customer,
             $orderId,
             $metadata,
-            $basket);
+            $basket,
+            $invoiceId,
+            $paymentReference
+        );
     }
 
     /**
      * Performs a Payout transaction and returns the resulting Payout resource.
      *
-     * @param float                $amount    The amount to charge.
-     * @param string               $currency  The currency of the amount.
-     * @param Payment              $payment   The payment object associated with the payout.
-     * @param string               $returnUrl The URL used to return to the shop if the process requires leaving it.
-     * @param Customer|string|null $customer  The customer associated with the payout.
-     * @param string|null          $orderId   A custom order id which can be set by the merchant.
+     * @param float                $amount           The amount to charge.
+     * @param string               $currency         The currency of the amount.
+     * @param Payment              $payment          The payment object associated with the payout.
+     * @param string               $returnUrl        The URL used to return to the shop if the process requires leaving it.
+     * @param Customer|string|null $customer         The customer associated with the payout.
+     * @param string|null          $orderId          A custom order id which can be set by the merchant.
      * @param null                 $metadata
      * @param null                 $basket
+     * @param string|null          $invoiceId        The external id of the invoice.
+     * @param string|null          $paymentReference A reference text for the payment.
      *
      * @return Payout The resulting object of the Payout resource.
      *
@@ -372,9 +394,14 @@ class PaymentService
         $customer = null,
         $orderId = null,
         $metadata = null,
-        $basket = null
+        $basket = null,
+        $invoiceId = null,
+        $paymentReference = null
     ): AbstractTransactionType {
-        $payout = (new Payout($amount, $currency, $returnUrl))->setOrderId($orderId);
+        $payout = (new Payout($amount, $currency, $returnUrl))
+            ->setOrderId($orderId)
+            ->setInvoiceId($invoiceId)
+            ->setPaymentReference($paymentReference);
         $payment->setPayout($payout)->setCustomer($customer)->setMetadata($metadata)->setBasket($basket);
         $this->resourceService->create($payout);
 
@@ -484,16 +511,17 @@ class PaymentService
      *
      * @param Payment|string $payment
      * @param string|null    $invoiceId
+     * @param string|null    $orderId
      *
      * @return Shipment Resulting Shipment object.
      *
      * @throws HeidelpayApiException
      * @throws RuntimeException
      */
-    public function ship($payment, $invoiceId = null): AbstractHeidelpayResource
+    public function ship($payment, $invoiceId = null, $orderId = null): AbstractHeidelpayResource
     {
         $shipment = new Shipment();
-        $shipment->setInvoiceId($invoiceId);
+        $shipment->setInvoiceId($invoiceId)->setOrderId($orderId);
         $this->resourceService->getPaymentResource($payment)->addShipment($shipment);
         $this->resourceService->create($shipment);
         return $shipment;