+++ src/Traits/CanAuthorizeWithCustomer.php
@@ -44,9 +44,11 @@ trait CanAuthorizeWithCustomer
      * @param Customer|string $customer
      * @param string|null     $orderId
      * @param Metadata|null   $metadata
-     * @param Basket|null     $basket   The Basket object corresponding to the payment.
-     *                                  The Basket object will be created automatically if it does not exist
-     *                                  yet (i.e. has no id).
+     * @param Basket|null     $basket           The Basket object corresponding to the payment.
+     *                                          The Basket object will be created automatically if it does not exist
+     *                                          yet (i.e. has no id).
+     * @param string|null     $invoiceId        The external id of the invoice.
+     * @param string|null     $paymentReference A reference text for the payment.
      *
      * @return Authorization
      *
@@ -60,7 +62,9 @@ trait CanAuthorizeWithCustomer
         $customer,
         $orderId = null,
         $metadata = null,
-        $basket = null
+        $basket = null,
+        $invoiceId = null,
+        $paymentReference = null
     ): Authorization {
         if ($this instanceof HeidelpayParentInterface) {
             return $this->getHeidelpayObject()->authorize(
@@ -71,7 +75,9 @@ trait CanAuthorizeWithCustomer
                 $customer,
                 $orderId,
                 $metadata,
-                $basket
+                $basket,
+                $invoiceId,
+                $paymentReference
             );
         }