+++ test/BasePaymentTest.php
@@ -56,10 +56,8 @@ class BasePaymentTest extends TestCase
     // thus can create a CreditCard via this SDK.
     // If the merchant is not certified to handle the CreditCard data SAQ-A applies
     // in which case the merchant has to embed our iFrame via JS (UIComponents).
-    const PRIVATE_KEY_SAQ_D = 's-priv-2a102ZMq3gV4I3zJ888J7RR6u75oqK3n';
-    const PUBLIC_KEY_SAQ_D  = 's-pub-2a10ifVINFAjpQJ9qW8jBe5OJPBx6Gxa';
-    const PRIVATE_KEY_SAQ_A = 's-priv-2a1095rIVXy4IrNFXG6yQiguSAqNjciC';
-    const PUBLIC_KEY_SAQ_A  = 's-pub-2a10xITCUtmO2FlTP8RKB3OhdnKI4RmU';
+    const PRIVATE_KEY = 's-priv-2a102ZMq3gV4I3zJ888J7RR6u75oqK3n';
+    const PUBLIC_KEY  = 's-pub-2a10ifVINFAjpQJ9qW8jBe5OJPBx6Gxa';
 
     /**
      * {@inheritDoc}
@@ -68,7 +66,7 @@ class BasePaymentTest extends TestCase
      */
     protected function setUp()
     {
-        $this->heidelpay = (new Heidelpay(self::PRIVATE_KEY_SAQ_D))
+        $this->heidelpay = (new Heidelpay(self::PRIVATE_KEY))
             ->setDebugHandler(new TestDebugHandler())->setDebugMode(true);
     }