+++ CHANGELOG.md
@@ -3,6 +3,35 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.2.0][1.2.2.0]
+
+### Fixed
+*   An issue resulting in an error when a B2B customer is fetched which does not have a function set.
+
+### Added
+*   Example to fetch all registered webhooks for the given key pair.
+*   Missing API response code for insurance already activated message after shipment.
+*   Giropay example
+*   Optional parameter orderId to ship call.
+
+### Changed
+*   Removed deletion of all webhooks prior to registering to webhook events.
+
+## [1.2.1.0][1.2.1.0]
+
+### Added
+*   Parameters `paymentReference` and `invoiceId` to `Authorization` and `Payout`.
+*   The SDK now supports the webhook event `payout`.
+*   Example for Flexipay direct.
+*   Email parameter to `Paypal` payment type.
+*   Error id to `HeidelpayApiException`.
+
+### Changed
+*   The webhook tests now cover all supported events.
+
+### Removed
+*   SAQ-A test due to lack of a corresponding key.
+
 ## [1.2.0.0][1.2.0.0]
 
 ### Changed
@@ -243,3 +272,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 [1.1.5.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.4.0..1.1.5.0
 [1.1.6.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.5.0..1.1.6.0
 [1.2.0.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.1.6.0..1.2.0.0
+[1.2.1.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.2.0.0..1.2.1.0
+[1.2.2.0]: https://github.com/heidelpay/heidelpayPHP/compare/1.2.1.0..1.2.2.0